requirements.txt
# Core FastAPI dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# LlamaIndex core (versão mais recente e estável)
llama-index-core>=0.10.0
llama-index-readers-file>=0.1.0
llama-index-llms-gemini>=0.1.0
llama-index-embeddings-gemini>=0.1.0

# Google AI
google-generativeai>=0.3.0

# Core utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=23.0.0

# Optional dependencies for scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Vector store (mais leve que ChromaDB)
faiss-cpu>=1.7.0

# Additional utilities
tenacity>=8.0.0
tiktoken>=0.4.0
